+++ CHANGES.md
@@ -1,7 +1,10 @@
 Change Log
 ==========
 
-### changes in 0.8-SNAPSHOT
+### changes in 0.9-SNAPSHOT
+
+
+### changes in 0.8
 
 * updated scala, sbt and associated dependencies
 * added def stringToIdentifierOption(s: String) to Identifier